wrap up fe for adding single correct mcq
